[Cleanup] Move classes in pinot-common to the correct package (#13478)
